Improve error handling in debug panel
Refactored the debug panel creation logic to only include an error list if there are errors present. This change ensures that the debug information presented to developers is more relevant and less cluttered, facilitating quicker identification and resolution of issues. Errors are now more prominently formatted and added directly to the content string when present.