Handle unlimited result set in dynamic content loader
Modified the dynamic.js to interpret a data-maxreturn value of '-1' as a request to load a maximum of 1000 items. This change provides a way to specify an "unlimited" number of results in a controlled manner, catering to use cases where a large result set is needed without exposing the system to unbounded queries.