Optimize Adobe metadata plugin scope
Removed unused `buildscript` variable from the global scope and reallocated its declaration within `handleMetadataTracking` function to enhance encapsulation and reduce clutter. This change improves the maintainability of the Adobe metadata plugin by ensuring that variables are defined closer to their actual usage. It also potentially reduces memory usage by not allocating memory for the variable when it's not needed.